deployment/requirements.txt
streamlit
tensorflow-cpu
Pillow
numpy
requests
streamlit_lottie
opencv-python-headless